// src/modules/ai/types.ts
export type MessageRole = "user" | "model";

export interface MessagePart {
    text: string;
}

export interface Message {
    role: MessageRole;
    parts: MessagePart[];
}

export interface ChatHistory extends Array<Message> {}


export interface GenerateConfig{
    temperature: number;
    topP: number;
    responseMimeType: string;
}

export interface ChatSettings{
    temperature: number;
    model: string;
    systemInstruction: string;
}